Rename the `utils` module to `helpers` for internal helpers in the library
